Encryption (#61)
* Encryption functional

* encryption differential files

* variable time encryptions added in the log file

* Preparation for the management of encryption parameters by ini

* Creation of the variable storing the chosen extension

* add class ini

* Create methods changeExtension

* choose encryption and save functional

* Abandonment of the ini file as obsolete

* Persistence of the value of the extension

* Finish

* Bug log encryption

* Bug name destination path file encrypted

* does not delete encrypted files

* Ok

+++ easySave/Models/job.cs
@@ -13,6 +13,7 @@ using System.Text;
 using Newtonsoft.Json;
 using System.Runtime.CompilerServices;
 using Newtonsoft.Json;
+using System.Diagnostics;
 
 [assembly: InternalsVisibleTo("easySave - Graphic")]
 
@@ -32,6 +33,7 @@ namespace easySave.Models
         string jsonStringLogProgress;
         string pathFileLogProgress = @"C:\EasySave\Log\logProgressSave.json";
         string pathfolderLog;
+
         /// <summary>
         /// Stores the job name 
         /// </summary>
@@ -55,7 +57,7 @@ namespace easySave.Models
         bool typeSave;
 
 
-
+        ///string encryptionExtension;
         #endregion
 
         #region properties
@@ -84,6 +86,12 @@ namespace easySave.Models
         /// Getter setter of the pathSource attribute
         /// </summary>
         public string PathDestination { get => pathDestination; set => pathDestination = value; }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        ///[JsonIgnore]
+        ///public string EncryptionExtension { get => encryptionExtension; set => encryptionExtension = value; }
 
         public string GetPathFileLogProgress()
         {
@@ -130,7 +138,7 @@ namespace easySave.Models
         /// Method to start a backup
         /// </summary>
         /// <returns>Return if the backup is well done</returns>
-        public bool copy()
+        public bool copy(string encryptionExtension = null)
         {
             bool confirmSave = false; //Confirmation of the backup execution - Set to false
 
@@ -148,13 +156,13 @@ namespace easySave.Models
                         destination.Delete(true); //Delete the directory
                     }
 
-                    copyComplete(source, destination); //Launch backup
+                    copyComplete(source, destination, encryptionExtension); //Launch backup
 
                     confirmSave = true; //Validate the backup
                 }
                 else //Differential
                 {
-                    copyDifferential(source, destination); //Launch backup
+                    copyDifferential(source, destination, encryptionExtension); //Launch backup
 
                     compareDelete(this.pathSource, this.pathDestination); //Delete non-existent files in the source
 
@@ -230,7 +238,7 @@ namespace easySave.Models
         /// </summary>
         /// <param name="source">Source DirectoryInfo</param>
         /// <param name="destination">Destination DirectoryInfo</param>
-        public void copyComplete(DirectoryInfo source, DirectoryInfo destination)
+        public void copyComplete(DirectoryInfo source, DirectoryInfo destination, string encryptionExtension = null)
         {
             //Cache directories before we start copying
             DirectoryInfo[] folders = source.GetDirectories(); 
@@ -249,8 +257,21 @@ namespace easySave.Models
                 
                 DateTime transferDelay = DateTime.Now;
 
-                //Copy the file to the target folder
-                file.CopyTo(Path.Combine(destination.FullName, file.Name), true);
+                if (encryptionExtension != null && encryptionExtension != "" && file.Extension == encryptionExtension)
+                {
+                    int delayEncryption = encryption(file, destination);
+
+                    logProgress.EncryptionTime = delayEncryption.ToString();
+
+                    logProgress.FileTarget = Path.Combine(destination.FullName, Path.GetFileNameWithoutExtension(destination.FullName) + ".cry");
+                }
+                else
+                {
+                    //Copy the file to the target folder
+                    file.CopyTo(Path.Combine(destination.FullName, file.Name), true);
+
+                    logProgress.EncryptionTime = "0";
+                }
 
                 TimeSpan timeSpan = DateTime.Now - transferDelay;
 
@@ -282,7 +303,7 @@ namespace easySave.Models
         /// </summary>
         /// <param name="source">Source DirectoryInfo</param>
         /// <param name="destination">Source DirectoryInfo</param>
-        public void copyDifferential(DirectoryInfo source, DirectoryInfo destination)
+        public void copyDifferential(DirectoryInfo source, DirectoryInfo destination, string encryptionExtension = null)
         {
             //Cache directories before we start copying
             DirectoryInfo[] folders = source.GetDirectories();
@@ -317,8 +338,22 @@ namespace easySave.Models
 
                             DateTime transferDelay = DateTime.Now;
 
-                            //Copy the file to the target folder
-                            file.CopyTo(Path.Combine(destination.FullName, file.Name), true);
+                            if (encryptionExtension != null && encryptionExtension != "" && file.Extension == encryptionExtension)
+                            {
+                                int delayEncryption = encryption(file, destination);
+
+                                logProgress.EncryptionTime = delayEncryption.ToString();
+
+                                logProgress.FileTarget = Path.Combine(destination.FullName, Path.GetFileNameWithoutExtension(destination.FullName) + ".cry");
+
+                            }
+                            else
+                            {
+                                //Copy the file to the target folder
+                                file.CopyTo(Path.Combine(destination.FullName, file.Name), true);
+
+                                logProgress.EncryptionTime = "0";
+                            }
 
                             TimeSpan timeSpan = DateTime.Now - transferDelay;
 
@@ -352,8 +387,22 @@ namespace easySave.Models
                     logProgress.FileSize = file.Length.ToString();
 
                     DateTime transferDelay = DateTime.Now;
-                    //Copy the file to the target folder only if it does not exist
-                    file.CopyTo(Path.Combine(destination.FullName, file.Name), false);
+
+                    if (encryptionExtension != null && encryptionExtension != "" && file.Extension == encryptionExtension)
+                    {
+                        int delayEncryption = encryption(file, destination);
+
+                        logProgress.EncryptionTime = delayEncryption.ToString();
+
+                        logProgress.FileTarget = Path.Combine(destination.FullName, Path.GetFileNameWithoutExtension(destination.FullName) + ".cry");
+                    }
+                    else
+                    {
+                        //Copy the file to the target folder
+                        file.CopyTo(Path.Combine(destination.FullName, file.Name), true);
+
+                        logProgress.EncryptionTime = "0";
+                    }
 
                     TimeSpan timeSpan = DateTime.Now - transferDelay;
 
@@ -420,7 +469,10 @@ namespace easySave.Models
                 //Try catch which will allow error handling if needed
                 try
                 {
-                    File.Delete(Path.Combine(destination, file));
+                    if (!file.Contains(".cry"))
+                    {
+                        File.Delete(Path.Combine(destination, file));
+                    }
                     //Console.WriteLine("File delete : " + file);
 
                 }
@@ -462,6 +514,21 @@ namespace easySave.Models
             }
 
             return verif;
+        }
+
+        public int encryption(FileInfo fileSource, DirectoryInfo destination)
+        {
+            string pathCryptoSoft = @"C:\Program Files (x86)\CryptoSoft\CryptoSoft.exe";
+
+            string path = destination.FullName + @"\" + Path.GetFileNameWithoutExtension(fileSource.FullName) + ".cry";
+
+            var e = Process.Start(pathCryptoSoft, "\"" + fileSource.FullName + "\"" + " " + "\"" + path + "\"");
+
+            e.WaitForExit();
+
+            int returnEncryption = e.ExitCode;
+
+            return returnEncryption;
         }
         #endregion
